+++ zombieSurvival/myFunctions.hpp
@@ -3,6 +3,6 @@
 #include <constants.hpp>
 
 /// angle form first cords to seconds
-double calculateAngle(sf::Vector2f firstPos, sf::Vector2f secondPos);
-double cordsDistance(sf::Vector2f firstPos, sf::Vector2f secondPos);
+float calculateAngle(sf::Vector2f firstPos, sf::Vector2f secondPos);
+float cordsDistance(sf::Vector2f firstPos, sf::Vector2f secondPos);
 #endif
